pyproject.toml
[project]
name = "rt-dlm"
version = "0.1.0"
description = "Real-Time Deep Learning Model - AGI Research Project by AuralithAI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AuralithAI", email = "contact@auralithai.com"}
]
keywords = [
    "deep-learning",
    "machine-learning",
    "jax",
    "agi",
    "transformer",
    "language-model",
    "research"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GPU :: NVIDIA CUDA :: 12",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    # Core ML/DL frameworks
    "dm_haiku>=0.0.13",
    "jax>=0.4.35",
    "jaxlib>=0.4.35",
    "optax>=0.2.3",
    
    # Data processing
    "scikit-learn>=1.6.1",
    "numpy>=1.26,<2.1",
    "tensorflow>=2.18.0",
    
    # Memory and retrieval
    "faiss-cpu>=1.10.0",
    
    # Security
    "cryptography>=42.0.0",
    
    # Safe model serialization
    "safetensors>=0.4.0",
    
    # Fairness
    "fairlearn>=0.10.0",
    
    # Visualization
    "matplotlib>=3.10.0",
    "rich>=13.9.4",
    
    # Web interfaces
    "fastapi>=0.115.8",
    "pydantic>=2.10.6",
    "gradio>=5.15.0",
    
    # NLP
    "nltk>=3.9.1",
    "requests>=2.32.3",
    
    # Hyperparameter optimization
    "optuna>=4.2.1",
    
    # Quantum computing
    "qiskit>=1.3.0",
    "qiskit-aer>=0.15.1",
    
    # Experiment tracking
    "wandb>=0.19.2",
    
    # Multi-modal processing
    "Pillow>=8.0,<11.0",
    "opencv-python>=4.10.0.84",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "scipy>=1.14.1",
    "moviepy>=2.1.1",
    
    # Knowledge graphs
    "networkx>=3.4.2",
    
    # Logging
    "loguru>=0.7.3",
    "psutil>=6.1.1",
    
    # Configuration
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    
    # Distributed computing
    "ray>=2.40.0",
    
    # Safety and alignment
    "transformers>=4.47.1",
    "torch>=2.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "flake8>=7.1.1",
    "pre-commit>=3.6.0",
    "bandit>=1.7.10",
]

docs = [
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=2.0.0",
]

monitoring = [
    "prometheus-client>=0.19.0",
]

export = [
    "tf2onnx>=1.16.1",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
]

all = [
    "rt-dlm[dev,docs,monitoring,export]",
]

[project.urls]
Homepage = "https://github.com/AuralithAI/RT-DLM"
Documentation = "https://github.com/AuralithAI/RT-DLM/tree/main/docs"
Repository = "https://github.com/AuralithAI/RT-DLM.git"
Issues = "https://github.com/AuralithAI/RT-DLM/issues"
Changelog = "https://github.com/AuralithAI/RT-DLM/blob/main/CHANGELOG.md"

[project.scripts]
rtdlm-train = "src.train:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

# =============================================================================
# Black configuration
# =============================================================================
[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | __pycache__
)/
'''

# =============================================================================
# isort configuration
# =============================================================================
[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".git", "__pycache__", "venv", ".venv"]

# =============================================================================
# mypy configuration
# =============================================================================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
no_strict_optional = true
show_error_codes = true
pretty = true
exclude = [
    "tests/",
    "docs/",
    "venv/",
    ".venv/",
]

# =============================================================================
# pytest configuration
# =============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["src/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# Coverage configuration
# =============================================================================
[tool.coverage.run]
source = ["src/core", "src/config", "src/modules"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "htmlcov"

# =============================================================================
# Bandit configuration
# =============================================================================
[tool.bandit]
exclude_dirs = ["src/tests", "tests", "venv", ".venv"]
skips = ["B101", "B311"]  # Skip assert and random warnings
